_cite/requirements.txt
manubot~=0.6
PyYAML~=6.0
diskcache~=5.6
rich~=13.6
python-dotenv~=0.21
google-search-results~=2.4
selenium
requests
beautifulsoup4


